requirements.txt
aiohttp==3.7.4.post0
pytest-asyncio==0.15.1
pytest~=6.2.4
allure-pytest
pydantic